Refactor filter components and update imports for consistency
