+++ frontend/app/(public)/layout.tsx
@@ -1,8 +1,8 @@
 // These styles apply to every route in the application
 import './styles.css';
 import type { Metadata } from 'next';
-import Navbar from '@/app/(public)/_ui/navbar';
-import Footer from '@/app/(public)/_ui/footer';
+import Navbar from '@/app/(public)/_ui/Navbar';
+import Footer from '@/app/(public)/_ui/Footer';
 
 export const metadata: Metadata = {
   title: 'Dr. GW Williams STUCO',
